+++ examples/basic-i18n/src/locales/messages.ts
@@ -1,0 +1,28 @@
+import { LOCALES } from './locales';
+
+export const messages = {
+  [LOCALES.ENGLISH]: {
+    homeTitle: 'Home',
+    aboutTitle: 'About',
+    currentLocale: 'Current locale',
+    defaultLocale: 'Default locale',
+    configuredLocales: 'Configured locales',
+    localeSwitcher: 'Locale Switcher',
+    basicLayout: 'BasicLayout',
+    userLayout: 'UserLayout',
+    notFound: 'Not Found',
+    userLogin: 'User Login',
+  },
+  [LOCALES.ZH_CN]: {
+    homeTitle: '首页',
+    aboutTitle: '关于',
+    currentLocale: '当前语言',
+    defaultLocale: '默认语言',
+    configuredLocales: '配置的语言',
+    localeSwitcher: '语言切换',
+    basicLayout: '主布局',
+    userLayout: '用户布局',
+    notFound: '未找到页面',
+    userLogin: '用户登录页'
+  },
+};