+++ examples/basic-i18n/src/pages/NotFound/index.tsx
@@ -1,0 +1,7 @@
+import { FormattedMessage } from 'react-intl';
+
+export default function NotFound() {
+  return (
+    <h2><FormattedMessage id="notFound" /></h2>
+  );
+}