+++ packages/build-mpa-config/src/generate/index.ts
@@ -13,19 +13,30 @@ function generatePageFiles(api: IPluginAPI, options: IGeneratorOptions): IGenera
   }
 
   const { context: { userConfig } } = api;
-  generator.generateRunAppFile(userConfig);
+  // 在分析运行时依赖的场景下，不能直接执行 generator
+  // 需要在分析完成并执行 disableRuntimePlugins 逻辑后，再执行生成
+  const generateTasks = [];
+  generateTasks.push(() => {
+    generator.generateRunAppFile(userConfig);
+  });
 
   // Do not modify the page entry when entryPath ends with app.ts
   if (isAppEntry) {
     return {
+      generator,
+      generateTasks,
       entryPath: pageEntry,
       runAppPath: generator.runAppPath,
       routesFilePath: generator.routesFilePath,
     };
   }
+  generateTasks.push(() => {
+    generator.generateEntryFile();
+  });
 
-  generator.generateEntryFile();
   return {
+    generator,
+    generateTasks,
     entryPath: generator.entryPath,
     runAppPath: generator.runAppPath,
     routesFilePath: generator.routesFilePath,